refactor: #135 use auth login common
