+++ client/src/components/hooks/useAuthLogin.js
@@ -1,13 +1,36 @@
-/**
- * Create `useAuthLogin` hook, same as `useWelcomeBack`. This can be called inside in `useWelcomeBack` and `useLoginModal`.
- * This hook should receive a `onAuthCompleted` callback, which `useWelcomeBack` requires after successfully validating User credential.
- * 
- */
+import { useCallback } from 'react';
+import { useDispatch } from 'react-redux';
+import { onBindingAuthInfo } from '@cd/actions/userActions';
+import { validateReturningUserSW } from '@cd/hooks/useServiceWorker';
 
-const useAuthLogin = ({ onAuthCompleted = undefined }) => {
-	const onAuthCredentialSuccess = () => {
-		// dispatch(onBindingAuthInfo({ publicKey, user: userDetails }, onCompleted: onAuthCompleted));
-	}
+const useAuthLogin = ({ onAuthCompleted = () => {} } = {}) => {
+	const dispatch = useDispatch();
+
+	const onAuthCredentialSuccess = useCallback(
+		(result) => {
+			const { publicKey, userDetails } = result;
+			dispatch(onBindingAuthInfo({ publicKey, user: userDetails }, onAuthCompleted));
+		},
+		[dispatch],
+	);
+
+	const validateUserCredential = useCallback(async (password) => {
+		if (!password) {
+			return undefined;
+		}
+
+		try {
+			const result = await validateReturningUserSW(password);
+
+			// Similar to useCreateUser
+			return result;
+		} catch (err) {
+			console.error(`🚀 ~ >> ~ err`, err);
+			return undefined;
+		}
+	}, []);
+
+	return { validateUserCredential, onAuthCredentialSuccess };
 }
 
 export default useAuthLogin;
