+++ client/src/components/web-extension/WelcomeBack/WelcomeBackPage.jsx
@@ -1,113 +1,23 @@
-import React, { useState, useCallback } from 'react';
+import React from 'react';
 import { useNavigate } from 'react-router-dom';
-import { useDispatch } from 'react-redux';
-import { Formik } from 'formik';
-import { Button, Form, FormControl } from 'react-bootstrap';
-import { lockAccount } from '@cd/actions/userActions';
-import messages from '@cd/shared/formMessages';
-import useWelcomeBack from './useWelcomeBack';
+import { AuthLogin } from '@cd/components/web-extension/Common/AuthLogin';
+
 import './WelcomeBack.scss';
 
-const onValidatePassword = (values) => {
-	const errors = {};
-
-	if (!values.password) {
-		errors.password = messages.passwordRequired;
-	}
-
-	return errors;
-};
-
 const WelcomeBackPage = () => {
-	const { onAuthCredentialSuccess, validateUserCredential } = useWelcomeBack();
-	const dispatch = useDispatch();
 	const navigate = useNavigate();
-	const [serverErrors, setServerErrors] = useState(undefined);
-	const onValidate = useCallback((values) => onValidatePassword(values), []);
-	const handleFormSubmit = useCallback(
-		async (values) => {
-			if (values.password) {
-				const result = await validateUserCredential(values.password);
 
-				if (!result) {
-					setServerErrors({ message: messages.passwordWrong });
-					return;
-				}
-
-				result.publicKey && onAuthCredentialSuccess(result);
-			}
-		},
-		[onAuthCredentialSuccess, validateUserCredential],
-	);
-
-	const onChangeHandler = useCallback(
-		(e, handler) => {
-			if (e?.target?.value && serverErrors) {
-				setServerErrors(undefined);
-			}
-			handler('password', e?.target?.value);
-		},
-		[serverErrors],
-	);
-
-	const onReset = useCallback(async () => {
-		dispatch(lockAccount());
-		navigate('/connectAccount');
-	}, [dispatch, navigate]);
+	const handleOnLoginSuccess = () => {
+		navigate('/');
+	}
 
 	return (
 		<section className="cd_we_page--root">
-			<div className="cd_we_create-wallet-layout--root">
-				<Formik
-					initialValues={{
-						password: '',
-					}}
-					validate={onValidate}
-					onSubmit={handleFormSubmit}
-				>
-					{({ errors, touched, setFieldValue, handleBlur, handleSubmit }) => {
-						return (
-							<div className="cd_we_create-wallet-layout--body">
-								<Form noValidate onSubmit={handleSubmit}>
-									<Form.Text>
-										<h2>Welcome Back!</h2>
-									</Form.Text>
-									<Form.Group className="mb-3" controlId="exampleForm.ControlInput1">
-										<Form.Label>Enter password</Form.Label>
-										<FormControl
-											onBlur={handleBlur}
-											onChange={(e) => onChangeHandler(e, setFieldValue)}
-											name="password"
-											type="password"
-											placeholder="Enter password"
-										/>
-										{errors.password && touched.password && (
-											<Form.Text className="invalid-feedback" id="passwordHelpBlock">
-												{errors.password}
-											</Form.Text>
-										)}
-									</Form.Group>
-									{serverErrors && (
-										<Form.Text className="invalid-feedback" id="passwordHelpBlock">
-											{serverErrors.message}
-										</Form.Text>
-									)}
-									<div className="cd_we_page--bottom">
-										<Button type="submit" className="cd_we_btn-next" disabled={false}>
-											Unlock
-										</Button>
-										<div className="cd_we_welcomeBack--bottom">
-											<Button onClick={onReset} variant="link">
-												Log in as another user?
-											</Button>
-										</div>
-									</div>
-								</Form>
-							</div>
-						);
-					}}
-				</Formik>
-			</div>
+			<h2>Welcome Back!</h2>
+			<AuthLogin 
+				isShowReset 
+				onLoginSuccess={handleOnLoginSuccess} 
+			/>
 		</section>
 	);
 };
